requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.2.0
Werkzeug==1.0.1
cssselect==1.1.0
lxml==4.5.2
numpy==1.19.1
